feat: Add custom alert

+++ client/src/components/Admin/ProjectsAdmin.jsx
@@ -1,14 +1,15 @@
 import React, { Fragment, useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { useLocation } from 'react-router-dom';
+import { useSelector, useDispatch } from 'react-redux';
+import { displayAlert } from '../../utils/alert';
 import axios from 'axios';
 
 const ProjectsAdmin = () => {
   const [ searchString, setSearchString ] = useState('');
   const [ projects, setProjects ] = useState(null);
   const { loading, user } = useSelector(state => state.auth);
-  const location = useLocation();
-
+  const dispatch = useDispatch();;
+  
+  // Load current user's projects
   useEffect(() => {
       const fetchProjects = async () => {
         const id = user ? user._id : 'none'
@@ -23,6 +24,7 @@ const ProjectsAdmin = () => {
       fetchProjects();
   }, []);
 
+  
   const redirect = (page, projectId=null) => {
     if (page === 'add') {
         window.location.href = '/admin/projects/add';
@@ -32,7 +34,7 @@ const ProjectsAdmin = () => {
     
   };
 
-  
+
   const handleSearchChange = (e) => {
     const value = e.target.value;
     setSearchString(value);
@@ -50,16 +52,22 @@ const ProjectsAdmin = () => {
 
   const handleDelete = async (id) => {
     try {
-      const res = await axios.delete(`http://127.0.0.1:5000/api/projects/${id}`);
+      
+      if (window.confirm('Are you sure you want to delete this project?')){
+          const res = await axios.delete(`http://127.0.0.1:5000/api/projects/${id}`);
 
-      //To Do: Replace with user feedback
-      alert(res.data.msg);
+          // dispatch alert
+          displayAlert(dispatch, res.data.msg, 'success');
+
+          setProjects(prev => prev.filter(project => project._id !== id));
+          
+      }
+      
 
     } catch (error) {
       console.log(error);
 
-      //To Do: Replace with user feedback
-      alert(error);
+      displayAlert(dispatch, error, 'danger');
     }
   } 
 
@@ -86,7 +94,13 @@ const ProjectsAdmin = () => {
             <div className='flex items-center justify-center border border-gray-300 p-5 pt-10 rounded-lg min-h-50' id='project-list'>
               <ul className='w-full' >
                 {
-                  projects?.map(project => (
+                  projects === null || projects.msg?
+                   <div className='text-gray-300 text-center flex flex-col justify-center items-center'>
+                  <p className="text-gray-400 text-2xl font-bold">No projects</p>
+                  <i className="fa-solid fa-wrench p-10 text-4xl text-gray-400"></i>
+                </div>
+                  :
+                  projects.map(project => (
                     <li key={project._id} className='p-3 border-b border-gray-300 w-full flex justify-between items-center bg-linear-to-b from-gray-500 to-gray-700 mb-2 rounded-lg transition duration-300 cursor-pointer'>
                       <span className='p-2 text-sm'>{ project.title }</span>
                       <div className='flex gap-2'>
@@ -99,6 +113,7 @@ const ProjectsAdmin = () => {
                       </div>
                     </li>
                   ))
+                 
                 }           
               </ul>
             </div>
